refactor: Nest financial-commitment-reports tools.
